requirements.txt
pandas
psycopg2       
sqlalchemy
python-dotenv